Edit Champion statistics Entities, work on AddChampionStatisticsData, add StatRune icons in wwwroot folder

+++ Data/LoLStats.Data.Models/ChampionStarterItems.cs
@@ -17,8 +17,8 @@
 
         public virtual Champion Champion { get; set; }
 
-        public decimal PickRate { get; set; }
+        public int PickRate { get; set; }
 
-        public decimal WinRate { get; set; }
+        public int WinRate { get; set; }
     }
 }
